# Implementing Unarchive and Empty Trash buttons. Some standardization.
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@16408 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ administrator/components/com_contact/views/contacts/view.html.php
@@ -58,23 +58,42 @@ class ContactViewContacts extends JView
 	 */
 	protected function addToolbar()
 	{
-		$state = $this->get('state');
+		require_once JPATH_COMPONENT.'/helpers/contact.php';
+		$canDo	= ContactHelper::getActions($this->state->get('filter.category_id'));
+
 		JToolBarHelper::title(JText::_('COM_CONTACT_MANAGER_CONTACTS'), 'generic.png');
-		JToolBarHelper::addNew('contact.edit', 'JTOOLBAR_NEW');
-		JToolBarHelper::editList('contact.edit','JTOOLBAR_EDIT');
-		JToolBarHelper::divider();
-		JToolBarHelper::publish('contacts.publish','JTOOLBAR_PUBLISH');
-		JToolBarHelper::unpublish('contacts.unpublish','JTOOLBAR_UNPUBLISH');
-		JToolBarHelper::divider();
-		JToolBarHelper::archiveList('contacts.archive','JTOOLBAR_ARCHIVE');
-		if ($state->get('filter.published') == -2) {
+		
+		if ($canDo->get('core.create')) {
+			JToolBarHelper::addNew('contact.add','JTOOLBAR_NEW');
+		}
+		if ($canDo->get('core.edit')) {
+			JToolBarHelper::editList('contact.edit','JTOOLBAR_EDIT');
+		}
+		if ($canDo->get('core.edit.state')) {
+			if ($this->state->get('filter.published') != 2){
+				JToolBarHelper::divider();
+				JToolBarHelper::custom('contacts.publish', 'publish.png', 'publish_f2.png','JTOOLBAR_PUBLISH', true);
+				JToolBarHelper::custom('contacts.unpublish', 'unpublish.png', 'unpublish_f2.png', 'JTOOLBAR_UNPUBLISH', true);
+			}
+			if ($this->state->get('filter.published') != -1 ) {
+				JToolBarHelper::divider();
+				if ($this->state->get('filter.published') != 2) {
+					JToolBarHelper::archiveList('contacts.archive','JTOOLBAR_ARCHIVE');
+				}
+				else if ($this->state->get('filter.published') == 2) {
+					JToolBarHelper::unarchiveList('contacts.publish', 'JTOOLBAR_UNARCHIVE');
+				}
+			}	
+		}
+		if ($this->state->get('filter.published') == -2 && $canDo->get('core.delete')) {
 			JToolBarHelper::deleteList('', 'contacts.delete','JTOOLBAR_EMPTY_TRASH');
-		} else {
+		} else if ($canDo->get('core.edit.state')) {
 			JToolBarHelper::trash('contacts.trash','JTOOLBAR_TRASH');
 		}
-		JToolBarHelper::divider();
-
-		JToolBarHelper::preferences('com_contact');
+		if ($canDo->get('core.admin')) {
+			JToolBarHelper::divider();
+			JToolBarHelper::preferences('com_contact');
+		}
 		JToolBarHelper::divider();
 		JToolBarHelper::help('screen.contact','JTOOLBAR_HELP');
 	}
